refactor: split background job handlers
