+++ src/egui_app/controller/background_jobs/similarity.rs
@@ -1,0 +1,8 @@
+use super::*;
+
+pub(super) fn handle_similarity_prepared(
+    controller: &mut EguiController,
+    message: jobs::SimilarityPrepResult,
+) {
+    controller.handle_similarity_prep_result(message);
+}
